+++ packages/components/badge/index.js
@@ -1,0 +1,3 @@
+import Badge from './src/badge';
+
+export { Badge };