+++ packages/components/cell/index.js
@@ -1,0 +1,4 @@
+import CellGroup from './src/cellgroup';
+import Cell from './src/cell';
+
+export { CellGroup, Cell };