+++ packages/components/cell/src/cellgroup.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="b--cell-group">
+    <slot></slot>
+  </div>
+</template>
+<script>
+
+export default {
+  name: 'b-cell-group'
+}
+</script>
+
+<style lang="less">
+
+</style>