+++ packages/index.js
@@ -4,6 +4,8 @@ import { Button } from './components/button';
 import { Icon } from './components/icon';
 import { Flexbox, FlexboxItem } from './components/flexbox';
 import { Grid, GridItem } from './components/grid';
+import { Badge } from './components/badge';
+import { CellGroup, Cell } from './components/cell';
 
 const install = function(Vue) {
   Vue.component(Button.name, Button);
@@ -12,6 +14,9 @@ const install = function(Vue) {
   Vue.component(FlexboxItem.name, FlexboxItem);
 	Vue.component(Grid.name, Grid);
   Vue.component(GridItem.name, GridItem);
+  Vue.component(Badge.name, Badge);
+  Vue.component(CellGroup.name, CellGroup);
+  Vue.component(Cell.name, Cell);
 }
 
 if (typeof window !== 'undefined' && window.Vue) {